git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/csl-sol210-branch@88075 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/testsuite/objc.dg/proto-lossage-4.m
@@ -7,25 +7,25 @@
 typedef struct objc_object { struct objc_class *class_pointer; } *id;
 
 @protocol Proto
-- (long)someValue;
+- (int)someValue;
 @end
 
 @interface Obj
-- (long)anotherValue;
+- (int)anotherValue;
 @end
 
-long foo(void) {
-  long receiver = 2;
+int foo(void) {
+  int receiver = 2;
   Obj *objrcvr;
   Obj <Proto> *objrcvr2;
 
-  receiver += [receiver someValue]; /* { dg-warning "invalid receiver type .long int." } */
+  receiver += [receiver someValue]; /* { dg-warning "invalid receiver type .int( )?." } */
 /* { dg-warning "Messages without a matching method signature" "" { target *-*-* } 22 } */
 /* { dg-warning "will be assumed to return .id. and accept" "" { target *-*-* } 22 } */
 /* { dg-warning ".\.\.\.. as arguments" "" { target *-*-* } 22 } */
 /* { dg-warning "assignment makes integer from pointer without a cast" "" { target *-*-* } 22 } */
 
-  receiver += [receiver anotherValue]; /* { dg-warning "invalid receiver type .long int." } */
+  receiver += [receiver anotherValue]; /* { dg-warning "invalid receiver type .int( )?." } */
 /* { dg-warning "assignment makes integer from pointer without a cast" "" { target *-*-* } 28 } */
   
   receiver += [(Obj *)receiver someValue]; /* { dg-warning ".Obj. may not respond to .\\-someValue." } */
